requirements.txt
# Core Django Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0
pymongo==4.6.0
mongoengine==0.27.0

# Authentication & Security
PyJWT==2.8.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
requests==2.31.0
Pillow==10.0.1

# AI Integration
openai==1.3.0

# Django Extensions
django-extensions==3.2.3
django-filter==23.4
drf-spectacular==0.26.5

# Redis (Optional - for caching)
redis==5.0.1
django-redis==5.4.0

# Production optimizations
python-dotenv==1.0.0

# AWS Services
boto3==1.34.0
django-storages==1.14.2
django-ses==3.5.0

# AI Integration (for local development)
openai==1.3.0